Stock transfer implemented

+++ OperationMonitoring/Models/StorageHistory.cs
@@ -12,7 +12,7 @@ namespace OperationMonitoring.Models
     {
         public int Id { get; set; }
         public virtual Stock Stock { get; set; }
-        public int? Amount { get; set; }
+        public double? Amount { get; set; }
         public virtual Storage StorageTo { get; set; }
         public virtual HistoryType HistoryType { get; set; }
         public DateTime Date { get; set; }
